some changes in my custom scripts
wallpaper now uses per-screen offset for the text, i.e. even if the wp
is 16:9 and the screen is 16:10 (e.g), the text can be set to an offset
in order to start within the visible parts of the wallpaper.

shortcuts overview has now a timer backend to dynamically change the
clock etc.